requirements.txt
# https://github.com/WolfgangFahl/gremlin-python-tutorial/issues/1
futures
# https://pypi.org/project/gremlinpython/
gremlinpython
# python tests
pytest
# yaml parsing
pyyaml
